requirements.txt
pycord
python_dotenv